I'm getting rid of the temp file, put it back if needed
the number of td records is the same either way (68), but the number of opt
records actually increases from 214 to 218 without the temp file. something is
clearly changing, in addition to all of the new stereochemistry errors, but Matt
thinks the round trip shouldn't affect anything. See
https://openforcefieldgroup.slack.com/archives/CFA4NL63E/p1690378220840189?thread_ts=1690376827.598449&cid=CFA4NL63E
and surrounding thread