try msm for torsions but looks like a unit issue
qubekit seems to assume that the hessian from qcarchive is in atomic
units (Ht/bohr^2) and then converts to kcal/mol/A^2

the conformer coordinates were definitely in angstrom, at least according to the
pint quantity, so the R^2 term should definitely be A^2. then, if the unit
vectors have units of A, you end up with a crazy A^3 kcal/mol unit, so it seems
the unit vectors and eigenvectors must be unitless. then, also assuming that the
eigenvalues have the same units as the partial hessian (kcal/mol/A^2 after
conversion) gives A^2 * kcal/mol/A^2 = kcal/mol

but the resulting values are roughly 2 orders of magnitude greater than those in
Sage (~3 vs ~443, for example)