queue stuff was premature, faster without it
I forgot that I would have to loop over the results of the pool.imap, so it's
very natural to do the insertions as fragments are received from the pool. the
other way wouldn't work without substantial restructuring because you can't
pickle a sqlite connection (meaning a Store cannot be pickled to be shared
between threads). the actual SMILES processing and fragmentation should be the
expensive step that I want to be shared across threads anyway, not the database
insertion